리뷰 작성 레디스 추가 작업

+++ src/main/java/com/example/reviceservice/domain/dto/ReviewResponse.java
@@ -11,7 +11,7 @@ import java.util.List;
 @AllArgsConstructor
 public class ReviewResponse {
     private long totalCount;
-    private double score;
+    private float score;
     private int cursor;
     private List<ReviewDTO> reviews;
